Create BeagleMatrix.js
This program features a game where you can move a cursor on 4 8x8 matrices with the main objective of acquiring points and at the same time, avoiding the red dots or exceeding the bounds of the game. Every time you acquire a point, you will gain 10 points and the cursor will move faster.